requirements.txt
xapian-haystack
python-imaging
django-haystack
django-gravatar
django-extensions
django-tagging
sorl-thumbnail
django-smileys
south
django-pagination
django-sentry
recaptcha-client
